+++ src/scene/sprite-nine-slice/NineSliceSpritePipe.ts
