+++ src/scene/sprite-nine-slice/init.ts
@@ -1,4 +1,4 @@
 import { extensions } from '../../extensions/Extensions';
-import { NineSliceSpritePipe } from './NiceSliceSpritePipe';
+import { NineSliceSpritePipe } from './NineSliceSpritePipe';
 
 extensions.add(NineSliceSpritePipe);
